% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\name{cFlip}
\alias{cFlip}
\title{cFlip}
\usage{
cFlip(...)
}
\arguments{
\item{...}{arguments}
}
\value{
a \code{flip.object}
}
\description{
It concatenates flip objects
warning("More than one test statistic is imputed, only the first perms space will be stored.")
}
