\name{flip-internal}
\title{Internal flip objects}
\alias{.allpermutations}
\alias{.getAlternative}
\alias{.getNull}
\alias{.getXY}
\alias{.makeContrasts}
\alias{.npermutations}
\alias{.dependence.nptest}
\alias{.symmetry.nptest}
\alias{.PermSpaceMatchInput}
\alias{.setTail}
\description{Internal flip objects.}
\details{These are not to be called by the user.}
\keyword{internal}