% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\name{p.value}
\alias{p.value}
\alias{getFlip}
\title{getFlip}
\usage{
p.value(object)

getFlip(object, element)
}
\arguments{
\item{object}{a \code{flip.object}}

\item{element}{element to the returned}
}
\value{
\code{getFlip} retuns a vector with values of the element requested. \code{p.value} returns a vector of p-values.
}
\description{
\code{getFlip} returns a given element of a flip.object.
\code{p.value} returns the p-values of a \code{flip.object}.
}
