% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\docType{methods}
\name{hist}
\alias{hist}
\title{hist}
\usage{
\S4method{hist}{flip.object}(x, ...)
}
\arguments{
\item{x}{a flip.object}

\item{...}{additional arguments to be passed}
}
\description{
show the histogram of the distribution of the test statistics (computed under the null hypothesis).
}
