% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\name{p.value}
\alias{p.value}
\title{p.value}
\usage{
p.value(object, ...)
}
\arguments{
\item{object}{a flip-object}

\item{...}{additional arguments to be passed}
}
\value{
vector of p-values
}
\description{
returns a vector of p-values from a flip-object
}
