% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\name{result}
\alias{result}
\title{result}
\usage{
result(object, ...)
}
\arguments{
\item{object}{a flip-object}

\item{...}{additional arguments to be passed}
}
\description{
prints information about the flip-object
}
