\name{seeds}
\alias{seeds}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Seeds data
}
\description{
Famous seeds growing data from Pesarin, F. (2001) Multivariate Permutation Tests with Applications in Biostatistics. Wiley, New York.
}

\usage{
seeds
}

\format{
the data.frame contains the three columns:
grs, x, y
}

\keyword{htest}
