% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\name{size}
\alias{size}
\title{size}
\usage{
size(object, ...)
}
\arguments{
\item{object}{a flip-object}

\item{...}{additional arguments to be passed}
}
\description{
size of permutation space: number of permutations X number of variables
}
