% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\docType{methods}
\name{show,flip.object-method}
\alias{show,flip.object-method}
\title{show}
\usage{
\S4method{show}{flip.object}(object)
}
\arguments{
\item{object}{a flip-object}

\item{...}{additional arguments to be passed}
}
\section{Methods (by class)}{
\itemize{
\item \code{flip.object}: prints same as \code{summary}
}}

