% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-package.R
\docType{class}
\name{flip.object-class}
\alias{[,flip.object,ANY,ANY,ANY-method}
\alias{[,flip.object-method}
\alias{[[,flip.object-method}
\alias{arrayOrNULL}
\alias{arrayOrNULL-class}
\alias{cFlip}
\alias{data.frameOrNULL}
\alias{data.frameOrNULL-class}
\alias{draw}
\alias{flip.object-class}
\alias{getFlip}
\alias{hist,flip.object-method}
\alias{length,flip.object-method}
\alias{names,flip.object-method}
\alias{names<-,flip.object-method}
\alias{numericOrmatrixOrNULL}
\alias{numericOrmatrixOrNULL-class}
\alias{p.adjust,flip.object-method}
\alias{p.value}
\alias{p.value,flip.object-method}
\alias{plot}
\alias{plot,flip.object-method}
\alias{result}
\alias{result,flip.object-method}
\alias{show,flip.object-method}
\alias{size}
\alias{size,flip.object-method}
\alias{sort,flip.object-method}
\alias{summary,flip.object-method}
\title{Class "flip.object" (and related functions) for storing the result of the
function flip and flipMix}
\arguments{
\item{obj}{Any \code{flip-object}}

\item{element}{Character string of either slot names of \code{obj} (i.e.
\code{"res","call","permP","permT"},
\code{"permSpace","permY","tail","data","call.env"}), elements of slot
\code{obj@data} (e.g. \code{"Y","X","Z","Strata"}) or columns of
\code{obj@res} (usually \code{"Test","Stat",}\code{ "tail","p-value"}).
Specific uses: \code{"Adjust:"} returns all columns of adjusted p-values in
\code{obj@res}. Any among \code{"nperms","perms","B"} return the number of
permutation used.}

\item{...}{\code{flip-object}s to be concatenated.}
}
\description{
The class flip.object is the output of a call to \code{\link{flip}},
\code{\link{flipMix}}, \code{\link{npc}} and \code{\link{flip.adjust}}.  It
stores the information needed for various diagnostic plots.  Specific
functions to deal with these objects are also documented here.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("flip.object", ...)}.
}
\examples{

showClass("flip.object")

y=matrix(rnorm(50),10,5)
colnames(y)=c("X1","X2","Y1","Y2","Y3")
res=flip(y)

## Sort by p-values
sort(res)

## Selecting tests
res[1:2]
#same as
res["X"]
#different from (it selects tests having "1" or "2" in the name)
res[c("1","2")]

## Concatenates two flip-objects
cFlip(res[1:2],res[5])


#plotting results
plot(flip(y))

#Get any slot of the flip-object. eg the permutation space:
head(getFlip(res,"permT"))
#Get any element of the list obj@data. eg Y:
getFlip(res,"Y")
#Get any columns of the results table: obj@res. eg Statistic (choose among colnames(obj@res) ):
getFlip(res,"Stat")

}
\author{
livio finos
}
\keyword{classes}

