% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\docType{class}
\name{flip.object-class}
\alias{flip.object-class}
\title{Class flip.object}
\description{
The class flip.object is the output of a call to \code{\link{flip}}, \code{\link{flipMix}}, \code{\link{npc}}, \code{\link{flip.adjust}} etc.
}
