% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-object.R
\name{getFlip}
\alias{getFlip}
\title{getFlip}
\usage{
getFlip(obj, element)
}
\arguments{
\item{obj}{a \code{flip.object}}

\item{element}{element to the returned}
}
\value{
values of the element requested.
}
\description{
It returns a given element of a flip.object.
}
