% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-package.R
\name{seeds}
\alias{seeds}
\title{Seeds data}
\format{the data.frame contains the three columns: grs, x, y}
\description{
Famous seeds growing data from Pesarin, F. (2001) Multivariate Permutation
Tests with Applications in Biostatistics. Wiley, New York.
}
\keyword{htest}

